refactor(share-extension): separate lifecycle and business logic from touch bar items and delegates
